# trading-chats-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序API文档",
        "title": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/prompt-templates": {
            "get": {
                "description": "获取所有提示词模版列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "获取所有提示词模版",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的提示词模版",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "创建提示词模版",
                "parameters": [
                    {
                        "description": "提示词模版信息",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PromptTemplate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/prompt-templates/generate": {
            "post": {
                "description": "根据提示词模版和参数动态生成提示词",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "动态生成提示词",
                "parameters": [
                    {
                        "description": "生成提示词请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GenerateAIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/prompt-templates/tag": {
            "get": {
                "description": "根据标签获取提示词模版列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "根据标签获取提示词模版",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/prompt-templates/{id}": {
            "get": {
                "description": "根据ID获取提示词模版详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "根据ID获取提示词模版",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提示词模版ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新提示词模版信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "更新提示词模版",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提示词模版ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "提示词模版信息",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PromptTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除提示词模版",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提示词模版"
                ],
                "summary": "删除提示词模版",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提示词模版ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.GenerateAIRequest": {
            "type": "object",
            "properties": {
                "param1": {
                    "type": "string",
                    "example": "https://api.example.com/data1"
                },
                "param2": {
                    "type": "string",
                    "example": "https://api.example.com/data2"
                },
                "template_id": {
                    "type": "string",
                    "example": "69b4e08fbd4f24a37ff0a0cc"
                }
            }
        },
...
        "models.PromptTemplate": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {},
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "description": "股票、期货、期权等",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {}
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}